Update README.md
The guest house system project is software developed to maintain and manage the various guest house task and processes. Its main goals are to increase income, streamline operations, and eliminate human labor. The primary goal of guest house management is to maintain a steady influx of visitors and guests, as well as to promote the FUD guest house extensive range of services and USPs and how they benefit visiting customers through marketing campaigns.